feat: Next.js 서버 컴포넌트로 데이터 prefetch 구현
