+++ src/components/ui/LoadingSpinner.tsx
@@ -1,0 +1,26 @@
+import { cn } from '@/lib/cn'
+
+export interface LoadingSpinnerProps {
+	className?: string
+	size?: 'sm' | 'md' | 'lg'
+}
+
+export function LoadingSpinner({ className, size = 'md' }: LoadingSpinnerProps) {
+	const sizeClasses = {
+		sm: 'h-4 w-4',
+		md: 'h-8 w-8',
+		lg: 'h-12 w-12',
+	}
+
+	return (
+		<div className="flex items-center justify-center p-8">
+			<div
+				className={cn(
+					'border-muted border-t-primary animate-spin rounded-full border-2',
+					sizeClasses[size],
+					className,
+				)}
+			/>
+		</div>
+	)
+}
